% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\docType{data}
\name{climate}
\alias{climate}
\title{Snow_v_rain function}
\format{a data frame with 365 rows and 2 columns
\describe{
\item{precip} - {generated precipitation (inches) for each event}
\item{temp} - {generated temperatures (F) corresponding to each precipitaiton event }

}}
\usage{
climate
}
\description{
Generated data
}
\keyword{datasets}
