% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evap_waterstorage.R
\docType{data}
\name{evap_waterstorage}
\alias{evap_waterstorage}
\title{WatershedStorage function}
\format{A vector with 1 column with a 100 generated evaporation values
\describe{
\item{evaporation}-{Evaporation values in inches per year}

}}
\usage{
evap_waterstorage
}
\description{
Generated data for evaporation in inches per year
}
\keyword{datasets}
