% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxPrecip.R
\name{MaxPrecip}
\alias{MaxPrecip}
\title{Function to find the greatest recorded precipitation values in a daily rainfall dataset for a particular location during the given period}
\usage{
MaxPrecip(precip, station = "Cachuma")
}
\arguments{
\item{precip}{precipitation in inches/day}

\item{station}{describes the number or name of the precipitation gauge or station}
}
\value{
the greatest recorded rainfall value

the data of the greatest recorded rainfall value

the selected rainfall station
}
\description{
Function to find the greatest recorded precipitation values in a daily rainfall dataset for a particular location during the given period
}
