% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanPrecip.R
\name{MeanPrecip}
\alias{MeanPrecip}
\title{Function to find the average precipitation value in a daily rainfall dataset for a particular location during the given period}
\usage{
MeanPrecip(precip, station = "Cachuma")
}
\arguments{
\item{precip}{precipitation in inches/day}

\item{station}{describes the name of the precipitation gauge or station}
}
\value{
the average of the recorded rainfall values

the selected rainfall station
}
\description{
Function to find the average precipitation value in a daily rainfall dataset for a particular location during the given period
}
