% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip_waterstorage.R
\docType{data}
\name{precip_waterstorage}
\alias{precip_waterstorage}
\title{WatershedStorage function}
\format{A vector with 1 column with a 100 generated precipitation values
\describe{
\item{precipitation}-{Precipitation values in inches per year}

}}
\usage{
precip_waterstorage
}
\description{
Generated data for precipitation in inches per year
}
\keyword{datasets}
