% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoff_waterstorage.R
\docType{data}
\name{runoff_waterstorage}
\alias{runoff_waterstorage}
\title{WatershedStorage function}
\format{A vector with 1 column with a 100 generated runoff values
\describe{
\item{runoff}-{Runoff values in inches per year}

}}
\usage{
runoff_waterstorage
}
\description{
Generated data for runoff in inches per year
}
\keyword{datasets}
