% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow_v_rain.R
\name{snow_v_rain}
\alias{snow_v_rain}
\title{Volume of snow vs rain (inches)}
\usage{
snow_v_rain(climate)
}
\arguments{
\item{climate}{data frame including columns for precipitation (inches) and temperature (F) observations}
}
\value{
precip_summary a table with volume of precipitation that fell as rain vs snow and average temperature when precipitation fell as rain vs snow.
}
\description{
Takes precipitation (inches) and corresponding temperature (F) data and calculates the volume of precipitation that fell as rain vs snow
}
