% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WatershedStorage.R
\name{WatershedStorage}
\alias{WatershedStorage}
\title{Function describing water stored in a watershed/year using simplified inflow/outflow processes}
\usage{
WatershedStorage(precip, evap, runoff, watershed_size)
}
\arguments{
\item{precip}{precipitation in inches/year}

\item{evap}{evaporation in inches/year}

\item{runoff}{runoff in inches/year}

\item{watershed_size}{total area of watershed in sq miles}
}
\value{
storage the amount of water stored in the watershed in cubic ft/year
}
\description{
Function describing water stored in a watershed/year using simplified inflow/outflow processes
}
