% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip.R
\docType{data}
\name{precip}
\alias{precip}
\title{MaxPrecip function}
\format{A data frame with 365 rows and 2 columns
\describe{
\item{date}-{Generated dates from January 1, 1995 to December 31, 1995}
\item{station}-{Rainfall station (Cachuma, Gibraltar, Jameson, or San Marcos)}
\item{rainfall}-{Generated precipitation values in inches per day}

}}
\usage{
precip
}
\description{
Generated data
}
\keyword{datasets}
