% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WatershedStorage.R
\name{WatershedStorage}
\alias{WatershedStorage}
\title{Volume of Water Stored in Watershed}
\usage{
WatershedStorage(precip, evap, runoff, watershed_size)
}
\arguments{
\item{precip}{volume of precipitation in inches/year}

\item{evap}{volume of water evaporated in inches/year}

\item{runoff}{volume of runoff in inches/year}

\item{watershed_size}{total area of watershed in sq miles}
}
\value{
storage the volume of water stored in the watershed in cubic ft/year
}
\description{
Function describing water stored in a watershed/year using simplified inflow/outflow processes
}
